Add max_batch support
For long-running jobs with stringent cuts, it can be desirable to impose
a maximum number of Vegas batches to use before stopping a calculation.

+++ include/nuchic/Event.hh
@@ -79,8 +79,8 @@ class Event {
         double Weight() const;
         void Rotate(const std::array<double,9>&);
 
-        void SetBatch(const int&);
-        int Batch() const { return m_batch; }
+        void SetBatch(const size_t&);
+        size_t Batch() const { return m_batch; }
 
     private:
         static double AddEvents(double, const MatrixElementStruct&);
@@ -95,7 +95,7 @@ class Event {
         MatrixElementVec m_me;
         double m_vWgt{}, m_meWgt{};
         vParticles m_leptons{};
-        int m_batch;
+        size_t m_batch;
 };
 
 }
